Restructure generator
Ref. #252
